lab2/pyproject.toml
[project]
name = "lab2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "litestar>=2.18.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.4",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
